+++ public/mockServiceWorker.js
@@ -114,7 +114,9 @@ self.addEventListener('fetch', async function (event) {
       }
 
       const reqHeaders = serializeHeaders(request.headers)
-      const body = await request.text()
+      const body = request.headers.get('Content-Type')?.includes('json')
+        ? await request.json()
+        : await request.text()
 
       const rawClientMessage = await sendToClient(client, {
         type: 'REQUEST',
