+++ src/setupTests.extra-2.js
@@ -1,0 +1,7 @@
+import {server} from 'test/server'
+
+// enable API mocking in test runs using the same request handlers
+// as for the client-side mocking.
+beforeAll(() => server.listen())
+afterAll(() => server.close())
+afterEach(() => server.resetHandlers())
