+++ src/utils/__tests__/misc.final.js
@@ -1,0 +1,5 @@
+import {formatDate} from '../misc'
+
+test('formatDate formats the date to look nice', () => {
+  expect(formatDate(new Date('October 18, 1988'))).toBe('Oct 88')
+})
